Move `InFlightLimitLayer` into `layer` mod
